Implemented samples iterator. Code improvements.

+++ internal/telemetry/provider.go
@@ -10,29 +10,17 @@ type Provider interface {
 
 // Telemetry is the telemetry data from the drone sensors
 type Telemetry struct {
-	Timestamp time.Time
-
-	// Barometer
-	Altitude *float64
-
-	// Gyroscope
-	Roll  *float64
-	Pitch *float64
-	Yaw   *float64
-
-	// Accelerometer
-	AccelX *float64
-	AccelY *float64
-	AccelZ *float64
-
-	// GPS
-	Latitude    *float64
-	Longitude   *float64
-	GroundSpeed *int
-
-	// Magnetometer
-	GroundCourse *int
-
-	// Radio
-	RadioRSSI *int
+	Timestamp    time.Time // Timestamp of telemetry measurement
+	Altitude     *float64  // Barometric altitude in meters
+	Roll         *float64  // Roll angle in degrees
+	Pitch        *float64  // Pitch angle in degrees
+	Yaw          *float64  // Yaw angle in degrees
+	AccelX       *float64  // X-axis acceleration in m/s²
+	AccelY       *float64  // Y-axis acceleration in m/s²
+	AccelZ       *float64  // Z-axis acceleration in m/s²
+	Latitude     *float64  // GPS latitude in degrees
+	Longitude    *float64  // GPS longitude in degrees
+	GroundSpeed  *float64  // Ground speed in m/s
+	GroundCourse *float64  // Ground course (heading) in degrees
+	RadioRSSI    *int64    // Radio link RSSI in dBm
 }
